+++ examples/InvoiceFactoring/Constants.php
@@ -28,4 +28,3 @@ require_once __DIR__ . '/../Constants.php';
 define('EXAMPLE_PATH', __DIR__);
 define('EXAMPLE_URL', EXAMPLE_BASE_FOLDER . 'InvoiceFactoring');
 define('CONTROLLER_URL', EXAMPLE_URL . '/Controller.php');
-define('RETURN_CONTROLLER_URL', EXAMPLE_URL . '/ReturnController.php');