+++ examples/PayPal/ReturnController.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * This is the controller for the Invoice Factoring example.
- * It is called when the pay button on the index page is clicked.
+ * This is the return controller for the PayPal example.
+ * It is called when the client is redirected back to the shop from the PayPal page of the selected bank.
  *
  * Copyright (C) 2019 heidelpay GmbH
  *
@@ -33,14 +33,7 @@ require_once __DIR__ . '/../../../../autoload.php';
 use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
-use heidelpayPHP\Resources\Basket;
-use heidelpayPHP\Resources\Customer;
-use heidelpayPHP\Resources\EmbeddedResources\Address;
-use heidelpayPHP\Resources\EmbeddedResources\BasketItem;
-use heidelpayPHP\Resources\PaymentTypes\InvoiceFactoring;
-
-session_start();
-session_unset();
+use heidelpayPHP\Resources\TransactionTypes\Authorization;
 
 $clientMessage = 'Something went wrong. Please try again later.';
 $merchantMessage = 'Something went wrong. Please try again later.';
@@ -53,43 +46,31 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
     die();
 }
 
+session_start();
+
+// Retrieve the paymentId you remembered within the Controller
+if (!isset($_SESSION['PaymentId'])) {
+    redirect(FAILURE_URL, 'The payment id is missing.', $clientMessage);
+}
+$paymentId = $_SESSION['PaymentId'];
+
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
     // Create a heidelpay object using your private key and register a debug handler if you want to.
     $heidelpay = new Heidelpay(HEIDELPAY_PHP_PAYMENT_API_PRIVATE_KEY);
     $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    /** @var InvoiceFactoring $invoiceFactoring */
-    $invoiceFactoring = $heidelpay->createPaymentType(new InvoiceFactoring());
-
-    // A customer with matching addresses is mandatory for Invoice Factoring payment type
-    $customer = new Customer('Linda', 'Heideich');
-    $address  = new Address();
-    $address->setName('Linda Heideich')
-        ->setStreet('Vangerowstr. 18')
-        ->setCity('Heidelberg')
-        ->setZip('69155')
-        ->setCountry('DE');
-    $customer->setBirthDate('2000-02-12')->setBillingAddress($address)->setShippingAddress($address);
-
-    $orderId = str_replace(['0.', ' '], '', microtime(false));
-
-    // A Basket is mandatory for Invoice Factoring payment type
-    $basketItem = new BasketItem('Hat', 10.0, 10.0, 1);
-    $basket = new Basket($orderId, 10.0, 'EUR', [$basketItem]);
-
-    $transaction = $invoiceFactoring->charge(12.99, 'EUR', CONTROLLER_URL, $customer, $orderId, null, $basket);
-
-    // You'll need to remember the shortId to show it on the success or failure page
-    $_SESSION['ShortId'] = $transaction->getShortId();
-
-    // Redirect to the success or failure page depending on the state of the transaction
-    $payment = $transaction->getPayment();
-    if ($transaction->isSuccess()) {
+    // Redirect to success if the payment has been successfully completed.
+    $payment   = $heidelpay->fetchPayment($paymentId);
+    if ($payment->isCompleted() || $payment->isPending()) {
         redirect(SUCCESS_URL);
     }
 
     // Check the result message of the transaction to find out what went wrong.
+    $transaction = $payment->getAuthorization();
+    if (!$transaction instanceof Authorization) {
+        $transaction = $payment->getChargeByIndex(0);
+    }
     $merchantMessage = $transaction->getMessage()->getCustomer();
 } catch (HeidelpayApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
@@ -98,3 +79,4 @@ try {
     $merchantMessage = $e->getMessage();
 }
 redirect(FAILURE_URL, $merchantMessage, $clientMessage);
+