+++ CHANGELOG.md
@@ -3,6 +3,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.2.0][1.1.2.0]
+
+### Added
+*   Added 3ds flag to card payment type.
+*   Added example for invoice guaranteed payment type.
+*   Added example for PayPal payment type.
+*   Added example for the sepa direct debit guaranteed payment type.
+*   Added example for the deletion of all webhooks.
+
+### Fixed
+*   Added missing parameter `invoiceId` to `payment::ship` method.
+*   A problem which resulted in an error when the property BasketItem::AmountPerUnit is set to 0.0.
+
+### Changed
+*   Refactored implementation examples.
+*   Enabled fetching the payment via orderId if its id is not set.
+*   Changed the default values of Basket and BasketItem.
+*   Refactored updating resource properties.
+*   Adapt to small API changes.
+
 ## [1.1.1.0][1.1.1.0]
 
 ### Added
@@ -151,3 +171,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.0.2.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.0.1.0..1.0.2.0
 [1.1.0.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.0.2.0..1.1.0.0
 [1.1.1.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.0.0..1.1.1.0
+[1.1.2.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.1.0..1.1.2.0